fix: renamed components for EditAccount
